requirements.txt
openai-whisper
sounddevice
PyQt6
pyperclip
pyautogui
keyboard
numpy
torch
pywin32
